+++ packages/app/components/classmates.component.tsx
@@ -1,25 +1,66 @@
-import { Card, Text } from '@ui-kitten/components'
 import React from 'react'
-import { View, StyleSheet } from 'react-native'
+import { StyleSheet, ListRenderItemInfo } from 'react-native'
+import {
+  Divider,
+  List,
+  ListItem,
+  Icon,
+  IconProps,
+  Text,
+} from '@ui-kitten/components'
+import { fullName, guardians, sortByFirstName } from '../utils/peopleHelpers'
+import { useChild } from './childContext.component'
+import { useClassmates } from '@skolplattformen/api-hooks'
+import { ContactMenu } from './contactMenu.component'
+import { Classmate } from '@skolplattformen/embedded-api'
+import { translate } from '../utils/translation'
+
+interface ClassmatesProps {
+  setSelected: (value?: number | null) => void
+}
 
 export const Classmates = () => {
-  const cardHeader = () => {
-    return (
-      <View style={styles.topContainer}>
-        <Text category="h6">Klasslistan ej tillgänglig</Text>
-      </View>
-    )
-  }
+  const child = useChild()
 
+  const { data } = useClassmates(child)
+  const renderItemIcon = (props: IconProps) => (
+    <Icon {...props} name="people-outline" />
+  )
+  const [selected, setSelected] = React.useState()
+  const renderItem = ({ item, index }: ListRenderItemInfo<Classmate>) => (
+    <ListItem
+      accessibilityLabel={`${translate('classmates.child')} ${index + 1}`}
+      accessibilityHint={`${translate(
+        'classmates.contactsForGuardianFor'
+      )} ${fullName(item)}`}
+      title={fullName(item)}
+      onPress={() => setSelected(item)}
+      description={guardians(item.guardians)}
+      accessoryLeft={renderItemIcon}
+      accessoryRight={() => (
+        <ContactMenu
+          contact={item}
+          selected={item === selected}
+          setSelected={setSelected}
+        />
+      )}
+    />
+  )
   return (
-    <View style={styles.container}>
-      <Card header={cardHeader} style={styles.contentContainer}>
-        <Text>
-          Klasslista kan tyvärr inte visas längre. Vi jobbar på att lösa det,
-          och återkommer med information när vi vet mer.
+    <List
+      style={styles.container}
+      data={sortByFirstName(data)}
+      ItemSeparatorComponent={Divider}
+      ListHeaderComponent={
+        <Text category="h5" style={styles.listHeader}>
+          {data?.length
+            ? `${translate('classmates.class')} ${data[0].className}`
+            : translate('classmates.class')}
         </Text>
-      </Card>
-    </View>
+      }
+      renderItem={renderItem}
+      contentContainerStyle={styles.contentContainer}
+    />
   )
 }
 
@@ -38,7 +79,6 @@ const styles = StyleSheet.create({
     justifyContent: 'space-between',
   },
   listHeader: {
-    backgroundColor: '#fff',
     paddingTop: 10,
     paddingLeft: 15,
   },