+++ CHANGELOG.md
@@ -1,3 +1,75 @@
+# [1.45.0](https://github.com/kolplattformen/skolplattformen/compare/v1.44.0...v1.45.0) (2021-09-18)
+
+
+### Features
+
+* 🎸 Fler språk i schemat (uppd. curriculum) ([#476](https://github.com/kolplattformen/skolplattformen/issues/476)) ([6a30053](https://github.com/kolplattformen/skolplattformen/commit/6a3005330ccfffdc67c646ba212160cf6c87e6e2))
+
+# [1.44.0](https://github.com/kolplattformen/skolplattformen/compare/v1.43.2...v1.44.0) (2021-09-18)
+
+
+### Features
+
+* 🎸 Changed newsListItem display date to modified ([c1da658](https://github.com/kolplattformen/skolplattformen/commit/c1da6583881e6813eb68118fc41706fc3b65be9e))
+* 🎸 Modified used for child overview as well ([0ad8bfe](https://github.com/kolplattformen/skolplattformen/commit/0ad8bfe15c544f2fe34a7d7b2bc4284dec70a6b3))
+* 🎸 Updated embedded for correct sorting ([9efef8d](https://github.com/kolplattformen/skolplattformen/commit/9efef8dad9e44e80422feb94ff7c68b22f3be51f))
+
+## [1.43.2](https://github.com/kolplattformen/skolplattformen/compare/v1.43.1...v1.43.2) (2021-09-16)
+
+
+### Bug Fixes
+
+* 🐛 Småfixar i schemat ([2c770c0](https://github.com/kolplattformen/skolplattformen/commit/2c770c08fb6e8adea558ec0af1be77978af843e0))
+
+## [1.43.1](https://github.com/kolplattformen/skolplattformen/compare/v1.43.0...v1.43.1) (2021-09-16)
+
+
+### Bug Fixes
+
+* 🐛 Översättningar på menytabbarna ([#471](https://github.com/kolplattformen/skolplattformen/issues/471)) ([b7a732d](https://github.com/kolplattformen/skolplattformen/commit/b7a732d21b1a84c3a3a26b7eb680937384de4c93))
+
+# [1.43.0](https://github.com/kolplattformen/skolplattformen/compare/v1.42.2...v1.43.0) (2021-09-16)
+
+
+### Features
+
+* 🎸 Den nygamla klasslistan ([#462](https://github.com/kolplattformen/skolplattformen/issues/462)) ([f0d05a5](https://github.com/kolplattformen/skolplattformen/commit/f0d05a50000a2958db9c0abb9bd9d9d2546414e8)), closes [#441](https://github.com/kolplattformen/skolplattformen/issues/441)
+
+## [1.42.2](https://github.com/kolplattformen/skolplattformen/compare/v1.42.1...v1.42.2) (2021-09-15)
+
+
+### Bug Fixes
+
+* 🐛 Välja språk på Android var trasigt ([#467](https://github.com/kolplattformen/skolplattformen/issues/467)) ([d431d68](https://github.com/kolplattformen/skolplattformen/commit/d431d68f5cee6f3550ba7937aa9d177ab083baba))
+
+## [1.42.1](https://github.com/kolplattformen/skolplattformen/compare/v1.42.0...v1.42.1) (2021-09-14)
+
+
+### Bug Fixes
+
+* 🐛 Fixes long navigation names in tab bar ([fe51923](https://github.com/kolplattformen/skolplattformen/commit/fe5192324961f0c2a941c4fa182637dbf9199c0e))
+* decrease font weight to 500 ([69ff3ff](https://github.com/kolplattformen/skolplattformen/commit/69ff3ff0e1c65747a7aa6ef469187f06e2b0274c))
+* fontSize on tabBarLabel decreased ([36bc627](https://github.com/kolplattformen/skolplattformen/commit/36bc62752a97b62597f3a368152fd86e5a98212f))
+
+# [1.42.0](https://github.com/kolplattformen/skolplattformen/compare/v1.41.0...v1.42.0) (2021-09-14)
+
+
+### Bug Fixes
+
+* update colors to reach AAA ([0140984](https://github.com/kolplattformen/skolplattformen/commit/0140984918e13f63b6302dc34d3ce163b3b697df))
+
+
+### Features
+
+* 🎸 Fixes status page and bump embedded ([7b65652](https://github.com/kolplattformen/skolplattformen/commit/7b6565296e3895c13384af571e2c60bf0ee4e24a))
+
+# [1.41.0](https://github.com/kolplattformen/skolplattformen/compare/v1.40.1...v1.41.0) (2021-09-13)
+
+
+### Features
+
+* 🎸 release 2.0.0 for Android ([0dcb983](https://github.com/kolplattformen/skolplattformen/commit/0dcb98312faa6bc81c264e0f8e1ca810bc72b0c3))
+
 ## [1.40.1](https://github.com/kolplattformen/skolplattformen/compare/v1.40.0...v1.40.1) (2021-09-12)